remove geonames api call from frontend, make request backend route
